requirements.txt
Django==4.0.2
djangorestframework==3.13.1
python-dotenv==0.19.2
psycopg2==2.9.3
Pillow==9.0.1
django-cors-headers==3.11.0
